Split oneOffs.spec.js into individual test files
